+++ tests/test_annotate.py
@@ -123,5 +123,10 @@ class AnnotationGoldTest(CoverageTest):
         cov = coverage.Coverage()
         self.start_import_stop(cov, "mae")
         cov.annotate()
-        assert self.stdout() == "1\n2\n"
+        assert self.stdout() == (
+            "1\n" +
+            "2\n" +
+            "The annotate command will be removed in a future version.\n" +
+            "Get in touch if you still use it: ned@nedbatchelder.com\n"
+            )
         compare(gold_path("annotate/mae"), ".", "*,cover")